Revise enum tests for FPP autocoder
