+++ FppTest/test/enum/CMakeLists.txt
@@ -21,6 +21,7 @@ set(UT_MOD_DEPS
 
 # List all .cpp files as UT_SOURCE_FILES. Only the UT build is allowed.
 set(UT_SOURCE_FILES
+  "${CMAKE_CURRENT_LIST_DIR}/EnumTest.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/main.cpp"
 )
 register_fprime_ut()
