+++ FppTest/test/enum/README.md
@@ -1,9 +1,9 @@
-# Autocoders/Python/test/enum_xml
+# FppTest/test/enum
 
-This directory contains unit tests for the Python enum XML code generator.
+This directory contains unit tests for the FPP enum code generator.
 
 To use this directory, you must have installed F Prime, and you must be inside 
 the F Prime Python virtual environment.
 
-* To build the tests, run 'fprime-util build -ut'.
+* To build the tests, run 'fprime-util build --ut'.
 * To run the tests, run 'fprime-util check'.
